requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
neo4j==5.23.0
langchain==0.2.14
langchain-community==0.2.12
langchain-ollama==0.1.1
jinja2==3.1.4
